+++ src/components/task-progress/TaskCard.tsx
@@ -1,7 +1,7 @@
+import { useTasksAction } from '@/hooks/useTasksAction'
 import { TASK_PROGRESS_ID } from '@/constants'
 import type { Task } from '@/types'
 import TaskIcon from '../TaskIcon'
-import { useRecoilState } from 'recoil'
 import { tasksState } from '@/features/taskAtoms'
 
 interface TaskCardProps {
@@ -9,18 +9,8 @@ interface TaskCardProps {
 }
 
 const TaskCard = ({ task }: TaskCardProps): JSX.Element => {
+  const { moveTaskCard } = useTasksAction()
   const isStarted = task.progressOrder === TASK_PROGRESS_ID.NOT_STARTED
-  const [tasks, setTasks] = useRecoilState<Task[]>(tasksState)
-
-  const moveTaskCard = (taskId: number, directionNumber: 1 | -1): void => {
-    const updatedTasks: Task[] = tasks.map((task) =>
-      task.id === taskId ? { ...task, progressOrder: task.progressOrder + directionNumber } : task,
-    )
-
-    setTasks(updatedTasks)
-    // console.log(tasks)
-    // console.log(updatedTasks)
-  }
 
   return (
     <div className="bg-green-200 p-6 rounded-xl my-2 flex flex-col gap-y-2 text-xl relative">
