+++ src/hooks/useTasksAction.ts
@@ -1,0 +1,35 @@
+import { useRecoilState } from 'recoil'
+import type { Task } from '@/types'
+import { tasksState } from '@/features/taskAtoms'
+import { TASK_PROGRESS_ID } from '@/constants'
+
+interface useTaskActionType {
+  completeTask: (taskId: number) => void
+  moveTaskCard: (taskId: number, directionNumber: 1 | -1) => void
+}
+
+export const useTasksAction = (): useTaskActionType => {
+  const [tasks, setTasks] = useRecoilState<Task[]>(tasksState)
+
+  const completeTask = (taskId: number): void => {
+    const updatedTasks: Task[] = tasks.map((task) =>
+      task.id === taskId ? { ...task, progressOrder: TASK_PROGRESS_ID.COMPLETED } : task,
+    )
+    setTasks(updatedTasks)
+  }
+
+  const moveTaskCard = (taskId: number, directionNumber: 1 | -1): void => {
+    const updatedTasks: Task[] = tasks.map((task) =>
+      task.id === taskId ? { ...task, progressOrder: task.progressOrder + directionNumber } : task,
+    )
+
+    setTasks(updatedTasks)
+    // console.log(tasks)
+    // console.log(updatedTasks)
+  }
+
+  return {
+    completeTask,
+    moveTaskCard,
+  }
+}
